requirements.txt
Pillow==8.1.2
aiohttp==3.7.3
alpha-vantage==2.3.1
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
cycler==0.10.0
idna==2.10
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.3.4
multidict==5.1.0
nltk==3.5
numpy==1.20.1
pandas==1.2.2
pathlib==1.0.1
pip==21.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
regex==2020.11.13
requests==2.25.1
setuptools==53.0.0
six==1.15.0
soupsieve==2.2
tqdm==4.58.0
typing-extensions==3.7.4.3
urllib3==1.26.3
yarl==1.6.3
